# source/back-end/openapi-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Watson Airlines Customer Experience",
    "description": "This is a REST API for the Watson Airlines Customer Experience.",
    "contact": {
      "name": "Mauricio Badano",
      "email": "mauriciobadano@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:8080",
      "description": "Local Server"
    },
    {
      "url": "https://final.14r9zp2vmz3m.us-south.codeengine.appdomain.cloud",
      "description": "IBM Code Engine Deployment"
    }
  ],
  "paths": {
    "/airlines": {
      "get": {
        "summary": "Get every Airline",
        "description": "Retrieve every listed airline",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Airline"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/watsonairlines": {
      "get": {
        "summary": "Get Watson Airlines Info",
        "description": "Retrieve Information about Watson Airlines",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/flightsod/{originId}/{destinationId}": {
      "get": {
        "summary": "Get flight through origin and destination",
        "description": "Get flights based on user-defined origin and destination",
        "parameters": [
          {
            "name": "originId",
            "in": "path",
            "description": "ID of the origin airport",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "destinationId",
            "in": "path",
            "description": "ID of the destination airport",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Flight"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/flightsda/{date}/{airline}": {
      "get": {
        "summary": "Get flight through date and airline",
        "description": "Get flights based on user-defined date and airline",
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "description": "Date of the flight (e.g., YYYY-MM-DD)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "airline",
            "in": "path",
            "description": "ID of the airline",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Flight"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/flightsid/{id}": {
      "get": {
        "summary": "Get flight through id",
        "description": "Retrieve a specific flight through the object ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the flight",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Flight"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Flight": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "AIRLINE": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "required": {
                    "type": "boolean",
                    "example": true
                  }
                }
              },
              "FLIGHT_NUMBER": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "number"
                  },
                  "required": {
                    "type": "boolean",
                    "example": true
                  }
                }
              },
              "ORIGIN_AIRPORT": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "required": {
                    "type": "boolean",
                    "example": true
                  }
                }
              },
              "DESTINATION_AIRPORT": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "required": {
                    "type": "boolean",
                    "example": true
                  }
                }
              },
              "CANCELLED": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "boolean"
                  },
                  "default": {
                    "type": "boolean",
                    "example": false
                  }
                }
              },
              "DEPARTURE_DATE": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date-time"
                  },
                  "required": {
                    "type": "boolean",
                    "example": true
                  }
                }
              },
              "ARRIVAL_DATE": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date-time"
                  },
                  "required": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        }
      },
      "Airport": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "IATA_CODE": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "required": {
                    "type": "boolean",
                    "example": true
                  }
                }
              },
              "AIRPORT": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "required": {
                    "type": "boolean",
                    "example": true
                  }
                }
              },
              "CITY": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "required": {
                    "type": "boolean",
                    "example": true
                  }
                }
              },
              "STATE": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "COUNTRY": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "required": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        }
      },
      "Airline": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "IATA_CODE": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "required": {
                    "type": "boolean",
                    "example": true
                  }
                }
              },
              "AIRLINE": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "required": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "security": []
}